feat(design): restructure home to the wireframe: systems tier + lab log
Implements the spec wireframe (docs/IDENTITY-AND-DESIGN-DIRECTION.md
section 5) on top of the command-center hero:

- Selected systems: 3 flagship proof cards in the spec format
  (problem -> built -> proves) with mono status/type badges, replacing
  the proof-surfaces + setup-cards + 7-card work wall; the remaining
  lanes compress into hairline "other lanes" rows
- From the lab: new section listing the 3 newest writeups with mono
  dates and tag chips; app/page.tsx becomes a thin server component
  that reads them from content/writeups at build time so the homepage
  cannot drift from the actual archive
- Homepage client component moves to components/home-page.tsx;
  flagshipSystems/workLanes content lives in data/content.ts
- Nav gains Lab; process section relabeled "How I work"
- CLAUDE.md structure/sections docs updated

Dead CSS from the removed sections (setup/work-card blocks) is left
for a dedicated cleanup pass; several of those selectors are comma-
shared with live components and need per-selector verification.
